+++ docs/development/api/changes.md
@@ -6,6 +6,8 @@
   converted to a boolean, set to true if the asset had a `status` of `archived`,
   and the timestamp values that were previously stored in the `archived`
   attribute have been moved to a new `last_archived` attribute.
+  - Note that the `last_archived` attribute has been deprecated in 4.x and will
+    be removed in 5.x.
 - Logs can now have a status of `abandoned`, in addition to the existing `done`
   and `pending`.